+++ flask_monitoringdashboard/core/config/parser.py
@@ -47,7 +47,7 @@ def parse_string(parser, header, arg_name, arg_value):
     :param parser: the parser to be used for parsing
     :param header: name of the header in the configuration file
     :param arg_name: name in the configuration file
-    :param arg_value: default value, the the value is not found
+    :param arg_value: default value, if the value is not found
     """
     env = get_environment_var(arg_name)
     arg_value = env if env else arg_value
@@ -55,6 +55,40 @@ def parse_string(parser, header, arg_name, arg_value):
         return parser.get(header, arg_name)
     return arg_value
 
+def parse_github_repo_string(parser, header, arg_name, arg_value):
+    """
+    Parse an argument from the given parser. If the argument is not specified, return the default
+    value
+    :param parser: the parser to be used for parsing
+    :param header: name of the header in the configuration file
+    :param arg_name: name in the configuration file
+    :param arg_value: default value, if the value is not found
+    """
+    env = get_environment_var(arg_name)
+    arg_value = env if env else arg_value
+    if arg_value in ["your_github_token", "your_repo_owner", "your_repo_name"]:
+        return None
+    if parser.has_option(header, arg_name):
+        return parser.get(header, arg_name)
+    return arg_value
+
+def parse_list(parser, header, arg_name, arg_value):
+    """
+    Parse an argument from the given parser. If the argument is not specified, return the default
+    value
+    :param parser: the parser to be used for parsing
+    :param header: name of the header in the configuration file
+    :param arg_name: name in the configuration file
+    :param arg_value: default value, if the value is not found
+    """
+    env = get_environment_var(arg_name)
+    arg_value = env if env else arg_value
+    if parser.has_option(header, arg_name):
+        str_val = parser.get(header, arg_name).replace(" ", "")
+        if not str_val:
+            return arg_value
+        return str_val.split(',')
+    return arg_value
 
 def parse_bool(parser, header, arg_name, arg_value):
     """
@@ -63,7 +97,7 @@ def parse_bool(parser, header, arg_name, arg_value):
     :param parser: the parser to be used for parsing
     :param header: name of the header in the configuration file
     :param arg_name: name in the configuration file
-    :param arg_value: default value, the the value is not found
+    :param arg_value: default value, if the value is not found
     """
     env = get_environment_var(arg_name)
     arg_value = env if env else arg_value
@@ -79,7 +113,7 @@ def parse_literal(parser, header, arg_name, arg_value):
     :param parser: the parser to be used for parsing
     :param header: name of the header in the configuration file
     :param arg_name: name in the configuration file
-    :param arg_value: default value, the the value is not found
+    :param arg_value: default value, if the value is not found
     """
     env = get_environment_var(arg_name)
     arg_value = ast.literal_eval(env) if env else arg_value