+++ flask_monitoringdashboard/database/exception_occurrence.py
@@ -51,6 +51,7 @@ def add_exception_occurrence(
 ):
     """
     Add a new ExceptionOccurrence record.
+    :return: The saved ExceptionOccurrence record
     """
     exception_occurrence = ExceptionOccurrence(
         request_id=request_id,
@@ -61,6 +62,7 @@ def add_exception_occurrence(
     )
     session.add(exception_occurrence)
     session.commit()
+    return exception_occurrence
 
 
 def count_grouped_exceptions(session: Session):
@@ -200,6 +202,55 @@ def get_exceptions_with_timestamps_and_stack_trace_id(
     return result
 
 
+def get_exception_group_page_number_by_endpoint(
+    session: Session, per_page: int, endpoint_id: int, stack_trace_snapshot_id: int
+):
+    """
+    Get the page number of an exception group that has occurred for a specific endpoint
+    :param session: session for the database
+    :param per_page: number of items per page
+    :param endpoint_id: the id of the endpoint
+    :param stack_trace_snapshot_id: the id of the stacktrace snapshot
+    :return: The page number of the exception group (int)
+    """
+    subquery = (
+        session.query(
+            ExceptionOccurrence.stack_trace_snapshot_id,
+            (func.row_number().over(order_by=desc(func.max(Request.time_requested))) - 1).label("row_index")
+        )
+        .join(Request, ExceptionOccurrence.request)
+        .join(Endpoint, Request.endpoint)
+        .join(ExceptionType, ExceptionOccurrence.exception_type)
+        .join(ExceptionMessage, ExceptionOccurrence.exception_msg)
+        .filter(Endpoint.id == endpoint_id)
+        .group_by(
+            ExceptionType.type,
+            ExceptionMessage.message,
+            ExceptionOccurrence.stack_trace_snapshot_id
+        )
+        .subquery()
+    )
+    result = (
+        session.query(
+            (subquery.c.row_index // per_page + 1).label("page")
+        )
+        .filter(subquery.c.stack_trace_snapshot_id == stack_trace_snapshot_id)
+        .scalar()
+    )
+
+    return int(result)
+
+
+def check_if_stack_trace_exists(session: Session, exc: BaseException, tb: Union[TracebackType, None]) -> bool:
+    """
+    Check if a stack_trace_snapshot already exists in the database.
+    """
+
+    hashed_trace = hash_stack_trace(exc, tb)
+    stacktrace = get_stack_trace_by_hash(session, hashed_trace)
+    return stacktrace is not None
+
+
 def save_exception_occurence_to_db(
     request_id: int,
     session: Session,
@@ -215,8 +266,10 @@ def save_exception_occurence_to_db(
     existing_trace = get_stack_trace_by_hash(session, hashed_trace)
 
     if existing_trace:
+        is_new_group = False
         trace_id = int(existing_trace.id)
     else:
+        is_new_group = True
         trace_id = add_stack_trace_snapshot(session, hashed_trace)
         idx = 0
         while tb:
@@ -251,6 +304,7 @@ def save_exception_occurence_to_db(
 
     exc_msg_id = add_exception_message(session, str(exc))
     exc_type_id = add_exception_type(session, typ.__name__)
-    add_exception_occurrence(
+    exception_occurrence = add_exception_occurrence(
         session, request_id, trace_id, exc_type_id, exc_msg_id, is_user_captured
     )
+    return exception_occurrence.request.endpoint_id, trace_id, is_new_group