+++ flask_monitoringdashboard/views/deployment.py
@@ -10,7 +10,6 @@ from flask_monitoringdashboard.core.utils import get_details
 from flask_monitoringdashboard.database import session_scope
 
 
-
 @blueprint.route('/api/deploy_details')
 @secure
 def deploy_details():
@@ -43,5 +42,34 @@ def deploy_config():
             'outlier_detection_constant': config.outlier_detection_constant,
             'timezone': str(config.timezone),
             'colors': config.colors,
+        }
+    )
+
+
+@blueprint.route('/api/deploy_alert_config')
+@secure
+def deploy_alert_config():
+    """
+    :return: A JSON-object with alert configuration details
+    """
+    post_to_back_if_telemetry_enabled(**{'name': 'deploy_alert_config'})
+
+    return jsonify(
+        {
+            'alert_enabled': config.alert_enabled,
+            'alert_type': config.alert_type,
+            'email': {
+                'smtp_host': config.smtp_host,
+                'smtp_port': config.smtp_port,
+                'smtp_user': config.smtp_user,
+                'smtp_to': config.smtp_to
+            },
+            'issue': {
+                'repository_name': config.repository_name,
+                'repository_owner': config.repository_owner
+            },
+            'chat': {
+                'chat_platform': config.chat_platform
+            }
         }
     )